cmd name_change
