+++ cmd/tuf/staged/root.json
@@ -1,0 +1,1 @@
+{"signatures":[{"keyid":"c0bfbf5375ef05b30be57edb804d1c5443afecb5bbae62dc113b370deeeabfdd","method":"ed25519","sig":"39e5f0348fbc83b33670f5b51f2360f810df29c374ad5be73602ffbbec15535205b877a2c6f18e1b729cee11a84697d658400ee50203b434c9d8ab8849c98900"},{"keyid":"5db607c64d46eb6ec5f7b45d20e4958a47b48d86f410f6f3adb691ec441fc788","method":"ed25519","sig":"39e5f0348fbc83b33670f5b51f2360f810df29c374ad5be73602ffbbec15535205b877a2c6f18e1b729cee11a84697d658400ee50203b434c9d8ab8849c98900"}],"signed":{"_type":"root","consistent_snapshot":true,"expires":"2021-09-01T15:24:02Z","keys":{"5db607c64d46eb6ec5f7b45d20e4958a47b48d86f410f6f3adb691ec441fc788":{"keyid_hash_algorithms":["sha256","sha512"],"keytype":"ed25519","keyval":{"public":"24ca466adde42f09254cb93d403e57ff85860eb72af60f78cd17dcc550064241"},"scheme":"ed25519"},"c0bfbf5375ef05b30be57edb804d1c5443afecb5bbae62dc113b370deeeabfdd":{"keyid_hash_algorithms":["sha256","sha512"],"keytype":"ed25519","keyval":{"public":"24ca466adde42f09254cb93d403e57ff85860eb72af60f78cd17dcc550064241"},"scheme":"ed25519"}},"roles":{"root":{"keyids":["c0bfbf5375ef05b30be57edb804d1c5443afecb5bbae62dc113b370deeeabfdd","5db607c64d46eb6ec5f7b45d20e4958a47b48d86f410f6f3adb691ec441fc788"],"threshold":1}},"spec_version":"1.0","version":1}}
